+++ tripartite-ding/tripartite-ding-openapi/src/main/java/work/gaigeshen/tripartite/ding/openapi/parameters/process/DingProcessWorkRecordUpdateParameters.java
@@ -17,7 +17,7 @@ public class DingProcessWorkRecordUpdateParameters implements ClientParameters {
 
     public static class Request {
 
-        public Integer agentid;
+        public Long agentid;
 
         public String process_instance_id;
